requirements.txt
flask
flask-cors
python-dotenv
google-genai
sqlalchemy
gunicorn
flask-sqlalchemy
google-cloud-secret-manager
requests
google-auth